requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
aiohttp==3.9.5
google-cloud-storage==2.16.0
google-cloud-batch
pydantic-settings==2.2.1
